refactor: Change scoreMaximum type from Float to Int for consistency

+++ app/src/main/java/com/example/fit2081a1_yang_xingyu_33533563/data/model/entity/ScoreTypeDefinitionEntity.kt
@@ -10,5 +10,5 @@ import androidx.room.PrimaryKey
 data class ScoreTypeDefinitionEntity(
     @PrimaryKey(autoGenerate = true) val scoreDefId: Int = 0,
     val scoreTypeName: String,
-    val scoreMaximum: Float,
+    val scoreMaximum: Int,
 )
